+++ src/main/java/console/ConsoleController.java
@@ -1,14 +1,10 @@
 package console;
 
-import domain.Car;
-import domain.Cars;
-import domain.Name;
-import domain.RacingGame;
-import domain.RandomNumberGenerator;
-import domain.Record;
-import domain.TryCount;
+import domain.core.Car;
+import domain.core.RacingGame;
+import domain.core.RandomNumberGenerator;
+import domain.vo.Record;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class ConsoleController {
 
@@ -16,30 +12,14 @@ public class ConsoleController {
     private final ConsolePrinter printer = new ConsolePrinter();
 
     public void playGame() {
-        Cars cars = makeCars();
-        TryCount tryCount = readTryCount();
-        RacingGame racingGame = new RacingGame(new RandomNumberGenerator(), cars);
+        List<String> cars = reader.readCarNames();
+        int tryCount = reader.readTryCount();
+        RacingGame racingGame = new RacingGame(new RandomNumberGenerator(), cars, tryCount);
 
-        List<List<Record>> playRecords = racingGame.playGame(tryCount);
+        List<List<Record>> playRecords = racingGame.playGame();
         printer.printPlayRecords(playRecords);
 
         List<Car> winners = racingGame.winners();
         printer.printWinners(winners);
-    }
-
-    private Cars makeCars() {
-        List<String> inputNames = reader.readCarNames();
-        List<Car> cars = inputNames.stream()
-                .map(Name::new)
-                .map(Car::new)
-                .collect(Collectors.toList());
-
-        return new Cars(cars);
-    }
-
-    private TryCount readTryCount() {
-        int inputTryCount = reader.readTryCount();
-        
-        return new TryCount(inputTryCount);
     }
 }
