+++ src/main/java/domain/core/RacingGame.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.core;
 
+import domain.vo.Name;
+import domain.vo.Record;
+import domain.vo.TryCount;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,13 +11,23 @@ public class RacingGame {
 
     private final NumberGenerator numberGenerator;
     private final Cars cars;
+    private final TryCount tryCount;
 
-    public RacingGame(NumberGenerator numberGenerator, Cars cars) {
+    public RacingGame(NumberGenerator numberGenerator, List<String> carNames, int tryCount) {
         this.numberGenerator = numberGenerator;
-        this.cars = cars;
+        this.cars = toCars(carNames);
+        this.tryCount = new TryCount(tryCount);
     }
 
-    public List<List<Record>> playGame(TryCount tryCount) {
+    private Cars toCars(List<String> carNames) {
+        List<Car> cars = carNames.stream()
+                .map(Name::new)
+                .map(Car::new)
+                .collect(Collectors.toList());
+        return new Cars(cars);
+    }
+
+    public List<List<Record>> playGame() {
         List<List<Record>> records = new ArrayList<>();
 
         for (int i = 0; i < tryCount.value(); i++) {
