--- src/test/java/domain/TryCountTest.java
@@ -1,5 +1,0 @@
-package domain;
-
-class TryCountTest {
-
-}
