+++ src/test/java/domain/core/CarTest.java
@@ -1,7 +1,9 @@
-package domain;
+package domain.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.vo.Distance;
+import domain.vo.Name;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
@@ -24,7 +26,7 @@ class CarTest {
     @Test
     void 숫자가_4_이상이면_한_칸_이동한다() {
         // given
-        Car car = new Car(new Name("doggy"));
+        Car car = createCar("doggy");
         int number = 4;
 
         // when
@@ -37,7 +39,7 @@ class CarTest {
     @Test
     void 숫자가_3_이하이면_이동하지_않는다() {
         // given
-        Car car = new Car(new Name("doggy"));
+        Car car = createCar("doggy");
         int number = 3;
 
         // when
@@ -45,5 +47,9 @@ class CarTest {
 
         // then
         assertThat(result.value()).isEqualTo(0);
+    }
+
+    public Car createCar(String name) {
+        return new Car(new Name(name));
     }
 }
