+++ src/test/java/domain/core/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
